# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"NotificationItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"body",
					"read",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NotificationsBatchUpdateResult": {
				"properties": {
					"updated": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationDataResponse_NotificationsBatchUpdateResult_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationsBatchUpdateResult"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MarkMultipleNotificationsReadBody": {
				"properties": {
					"notificationIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"notificationIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotificationDataResponse_NotificationItem_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationItem"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.NewsCategory": {
				"type": "string",
				"enum": [
					"EARNINGS",
					"ANALYST",
					"FILING",
					"MERGER",
					"MACRO",
					"SECTOR",
					"GENERAL"
				]
			},
			"NewsCategory": {
				"$ref": "#/components/schemas/_36_Enums.NewsCategory"
			},
			"_36_Enums.NewsSentiment": {
				"type": "string",
				"enum": [
					"BULLISH",
					"BEARISH",
					"NEUTRAL"
				]
			},
			"NewsSentiment": {
				"$ref": "#/components/schemas/_36_Enums.NewsSentiment"
			},
			"NewsArticleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"headline": {
						"type": "string"
					},
					"summaryBullets": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"source": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string",
						"nullable": true
					},
					"publishedAt": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"$ref": "#/components/schemas/NewsCategory"
					},
					"sentiment": {
						"allOf": [
							{
								"$ref": "#/components/schemas/NewsSentiment"
							}
						],
						"nullable": true
					},
					"sentimentScore": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"relatedSymbols": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sector": {
						"type": "string",
						"nullable": true
					},
					"userContext": {
						"properties": {
							"inWatchlist": {
								"type": "boolean"
							},
							"inPortfolio": {
								"type": "boolean"
							}
						},
						"required": [
							"inWatchlist",
							"inPortfolio"
						],
						"type": "object"
					},
					"isRead": {
						"type": "boolean"
					},
					"isSaved": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"headline",
					"summaryBullets",
					"source",
					"url",
					"imageUrl",
					"publishedAt",
					"category",
					"sentiment",
					"sentimentScore",
					"relatedSymbols",
					"userContext",
					"isRead",
					"isSaved"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsFeedEnvelopeResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/NewsArticleResponse"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasMore": {
						"type": "boolean"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"nextCursor",
					"hasMore",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedNews": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/NewsArticleResponse"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasMore": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"nextCursor",
					"hasMore"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_PaginatedNews_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginatedNews"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsSummaryResponse": {
				"properties": {
					"portfolioNews": {
						"items": {
							"properties": {
								"isRead": {
									"type": "boolean"
								},
								"publishedAt": {
									"type": "string",
									"format": "date-time"
								},
								"sentiment": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NewsSentiment"
										}
									],
									"nullable": true
								},
								"symbol": {
									"type": "string",
									"nullable": true
								},
								"headline": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isRead",
								"publishedAt",
								"sentiment",
								"symbol",
								"headline",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"watchlistNews": {
						"items": {
							"properties": {
								"isRead": {
									"type": "boolean"
								},
								"publishedAt": {
									"type": "string",
									"format": "date-time"
								},
								"sentiment": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NewsSentiment"
										}
									],
									"nullable": true
								},
								"symbol": {
									"type": "string",
									"nullable": true
								},
								"headline": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isRead",
								"publishedAt",
								"sentiment",
								"symbol",
								"headline",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"marketHeadlines": {
						"items": {
							"properties": {
								"isRead": {
									"type": "boolean"
								},
								"publishedAt": {
									"type": "string",
									"format": "date-time"
								},
								"sentiment": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NewsSentiment"
										}
									],
									"nullable": true
								},
								"symbol": {
									"type": "string",
									"nullable": true
								},
								"headline": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isRead",
								"publishedAt",
								"sentiment",
								"symbol",
								"headline",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"unreadCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"portfolioNews",
					"watchlistNews",
					"marketHeadlines",
					"unreadCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_NewsSummaryResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NewsSummaryResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MarkReadResult": {
				"properties": {
					"updated": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"properties": {
								"isRead": {
									"type": "boolean"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"isRead",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"updated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_MarkReadResult_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/MarkReadResult"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MarkMultipleReadBody": {
				"properties": {
					"articleIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"articleIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticleActionState": {
				"properties": {
					"id": {
						"type": "string"
					},
					"isRead": {
						"type": "boolean"
					},
					"isSaved": {
						"type": "boolean"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewsDataResponse_ArticleActionState_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ArticleActionState"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SimpleNewsMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "stock-backend",
		"version": "1.0.0",
		"description": "Stock prediction, decision making and analytics platform.",
		"contact": {
			"name": "Hamza Ahmed"
		}
	},
	"paths": {
...
		"/api/v1/notifications/read-all": {
			"patch": {
				"operationId": "MarkAllAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationsBatchUpdateResult_"
								}
							}
						}
					}
				},
				"description": "Mark all unread notifications as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/notifications/read-multiple": {
			"patch": {
				"operationId": "MarkMultipleAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationsBatchUpdateResult_"
								}
							}
						}
					}
				},
				"description": "Mark multiple notifications as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MarkMultipleNotificationsReadBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/notifications/{id}/read": {
			"patch": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDataResponse_NotificationItem_"
								}
							}
						}
					}
				},
				"description": "Mark a single notification as read.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/news/feed": {
			"get": {
				"operationId": "GetNewsFeed",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsFeedEnvelopeResponse"
								}
							}
						}
					}
				},
				"description": "Get paginated financial news feed filtered by category, symbol, or portfolio/watchlist context.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewsCategory"
						}
					},
					{
						"in": "query",
						"name": "symbol",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"default": "all",
							"type": "string",
							"enum": [
								"portfolio",
								"watchlist",
								"all"
							]
						}
					}
				]
			}
		},
		"/api/v1/news/search": {
			"get": {
				"operationId": "SearchNews",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_PaginatedNews_"
								}
							}
						}
					}
				},
				"description": "Search news articles by text query, symbol, category, and date range.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "symbol",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewsCategory"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/news/summary": {
			"get": {
				"operationId": "GetNewsSummary",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_NewsSummaryResponse_"
								}
							}
						}
					}
				},
				"description": "Get personalized news summary: portfolio news, watchlist news, top market headlines, and unread count.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/news/saved": {
			"get": {
				"operationId": "GetSavedNews",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_PaginatedNews_"
								}
							}
						}
					}
				},
				"description": "Get user's saved/bookmarked news articles.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/news/symbol/{symbol}": {
			"get": {
				"operationId": "GetNewsBySymbol",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_PaginatedNews_"
								}
							}
						}
					}
				},
				"description": "Get news articles specific to a stock symbol.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/news/read-all": {
			"patch": {
				"operationId": "MarkAllAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_MarkReadResult_"
								}
							}
						}
					}
				},
				"description": "Mark all news articles as read for the user.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/news/read-multiple": {
			"patch": {
				"operationId": "MarkMultipleAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_MarkReadResult_"
								}
							}
						}
					}
				},
				"description": "Mark multiple news articles as read.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MarkMultipleReadBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/news/{id}/read": {
			"patch": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_ArticleActionState_"
								}
							}
						}
					}
				},
				"description": "Mark a single news article as read.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/news/{id}/save": {
			"post": {
				"operationId": "SaveArticle",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsDataResponse_ArticleActionState_"
								}
							}
						}
					}
				},
				"description": "Save/bookmark a news article.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "UnsaveArticle",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SimpleNewsMessageResponse"
								}
							}
						}
					}
				},
				"description": "Unsave/remove bookmark from a news article.",
				"tags": [
					"News"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}